[#1] Move argument parsing to its own module
